requirements.txt
langchain==0.2.7
langchain-openai==0.1.7
langchain-community==0.2.7
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
tabulate==0.9.0
langchain-groq==0.1.4
fastapi==0.115.0
uvicorn==0.30.6



